Product seller

+++ src/controllers/product.js
@@ -1,9 +1,12 @@
+/* eslint-disable consistent-return */
+/* eslint-disable max-len */
 const Sequelize = require('sequelize');
 const { cloudPathToFileName } = require('../helpers/converter');
 const { deleteImages } = require('../helpers/deleteArrayImages');
 const { deleteFile } = require('../helpers/fileHandler');
 const responseHandler = require('../helpers/responseHandler');
 const Category = require('../models/category');
+const ColorProduct = require('../models/colorProduct');
 const Product = require('../models/product');
 const ProductCategory = require('../models/productCategory');
 const ProductImage = require('../models/productImage');
@@ -32,7 +35,10 @@ exports.getAllProduct = async (req, res) => {
   const offset = (page - 1) * limit;
   const results = await Product.findAll({
     include: [
-      { model: ProductCategory },
+      {
+        model: ProductCategory,
+        attributes: ['id_category'],
+      },
       {
         model: ProductImage,
         attributes: ['image'],
@@ -53,6 +59,13 @@ exports.getAllProduct = async (req, res) => {
   });
   const count = await Product.count({
     where: {
+      name: {
+        [Sequelize.Op.like]: `%${search}%`,
+      },
+      price: {
+        [Sequelize.Op.gte]: minPrice,
+        [Sequelize.Op.lte]: maxPrice,
+      },
       is_deleted: 0,
     },
   });
@@ -64,7 +77,7 @@ exports.getAllProduct = async (req, res) => {
     currentPage: page,
     lastPage: last,
   };
-  return responseHandler(res, 200, 'List of products', results, pageInfo);
+  return responseHandler(res, 200, 'List of products', results.rows, pageInfo);
 };
 
 exports.getProductBySeller = async (req, res) => {
@@ -79,14 +92,72 @@ exports.getProductBySeller = async (req, res) => {
     if (!seller || seller.length < 1) {
       return responseHandler(res, 404, 'Seller not found');
     }
+    const { search = '' } = req.query;
+    let {
+      minPrice, maxPrice, limit, page,
+    } = req.query;
+    minPrice = parseInt(minPrice, 10) || 0;
+    maxPrice = parseInt(maxPrice, 10) || 100000000;
+    limit = parseInt(limit, 10) || 10;
+    page = parseInt(page, 10) || 1;
+    const dataName = ['search', 'minPrice', 'maxPrice'];
+    const data = { search, minPrice, maxPrice };
+    let url = `${APP_URL}/product/seller/${id}?`;
+    dataName.forEach((x) => {
+      if (req.query[x]) {
+        data[x] = req.query[x];
+        url = `${url}${x}=${data[x]}&`;
+      }
+    });
+    const offset = (page - 1) * limit;
     const product = await Product.findAll({
+      include: [
+        {
+          model: ProductCategory,
+          attributes: ['id_category'],
+        },
+        {
+          model: ProductImage,
+          attributes: ['image'],
+        },
+      ],
       where: {
+        name: {
+          [Sequelize.Op.like]: `%${search}%`,
+        },
+        price: {
+          [Sequelize.Op.gte]: minPrice,
+          [Sequelize.Op.lte]: maxPrice,
+        },
+        seller_id: id,
+        is_deleted: 0,
+      },
+      limit,
+      offset,
+    });
+    const count = await Product.count({
+      where: {
+        name: {
+          [Sequelize.Op.like]: `%${search}%`,
+        },
+        price: {
+          [Sequelize.Op.gte]: minPrice,
+          [Sequelize.Op.lte]: maxPrice,
+        },
         seller_id: id,
         is_deleted: 0,
       },
     });
+    const last = Math.ceil(count / limit);
+    const pageInfo = {
+      prev: page > 1 ? `${url}page=${page - 1}&limit=${limit}` : null,
+      next: page < last ? `${url}page=${page + 1}&limit=${limit}` : null,
+      totalData: count,
+      currentPage: page,
+      lastPage: last,
+    };
     if (product.length > 0) {
-      return responseHandler(res, 200, 'Product list', product, null);
+      return responseHandler(res, 200, 'Product list', product, pageInfo);
     }
     return responseHandler(res, 200, 'Seller has no product yet', null, null);
   } catch {
@@ -96,23 +167,22 @@ exports.getProductBySeller = async (req, res) => {
 
 exports.createProduct = async (req, res) => {
   try {
+    if (req.user.role !== 2) {
+      if (req.files) {
+        deleteImages(req.files);
+      }
+      return responseHandler(res, 401, 'You are not able to do this action');
+    }
     const listIdCategory = req.body.id_category.split(',');
     if (listIdCategory.length < 1) {
       return responseHandler(res, 400, 'Please enter at least 1 category', null, null);
     }
-    const seller = await User.findAll({
-      where: {
-        id: req.body.seller_id,
-        id_role: 2,
-      },
+    const dataProduct = {};
+    Object.keys(req.body).forEach((x) => {
+      dataProduct[x] = req.body[x];
     });
-    if (!seller || seller.length < 1) {
-      if (req.files) {
-        deleteImages(req.files);
-      }
-      return responseHandler(res, 404, 'Seller not found');
-    }
-    const product = await Product.create(req.body);
+    dataProduct.seller_id = req.user.id;
+    const product = await Product.create(dataProduct);
     const data = { id_product: product.dataValues.id };
     if (req.files) {
       req.files.forEach(async (pic) => {
@@ -173,9 +243,16 @@ exports.updateProduct = async (req, res) => {
       }
       return responseHandler(res, 404, 'Product not found', null, null);
     }
+    if (req.user.role !== 2 && req.user.id !== product.dataValues.seller_id) {
+      if (req.files) {
+        deleteImages(req.files);
+      }
+      return responseHandler(res, 401, 'You are not able to do this action');
+    }
     Object.keys(req.body).forEach((data) => {
       product[data] = req.body[data];
     });
+    product.seller_id = req.user.id;
     await product.save();
     const productImage = await ProductImage.findAll({
       where: {
@@ -206,23 +283,33 @@ exports.updateProduct = async (req, res) => {
 };
 
 exports.deleteProduct = async (req, res) => {
-  const { id } = req.params;
-  const product = await Product.findByPk(id);
-  if (product && product.dataValues.is_deleted === false) {
-    product.is_deleted = 1;
-    await product.save();
-    const productImage = await ProductImage.findAll({
-      where: {
-        id_product: id,
-      },
-    });
-    if (productImage) {
-      productImage.map(async (data) => {
-        await data.destroy();
-        deleteFile(cloudPathToFileName(data.image));
-      });
+  try {
+    const { id } = req.params;
+    const product = await Product.findByPk(id);
+    if (req.user.id !== product.dataValues.seller_id) {
+      return responseHandler(res, 401, 'You are not able to do this action');
     }
-    return responseHandler(res, 200, 'Product was deleted', null, null);
+    if (product && product.dataValues.is_deleted === false) {
+      product.is_deleted = 1;
+      await product.save();
+      const productImage = await ProductImage.findAll({
+        where: {
+          id_product: id,
+        },
+      });
+      if (productImage) {
+        productImage.map(async (data) => {
+          await data.destroy();
+          deleteFile(cloudPathToFileName(data.image));
+        });
+      }
+      return responseHandler(res, 200, 'Product was deleted', null, null);
+    }
+    return responseHandler(res, 404, 'Product not found', null, null);
+  } catch (e) {
+    if (req.files) {
+      deleteImages(req.files);
+    }
+    return responseHandler(res, 400, 'Can\'t delete', e, null);
   }
-  return responseHandler(res, 404, 'Product not found', null, null);
 };
